1) Delete work directories by default, use -savetemporary to disable 2) Can do just table building now 3) Simple checks to ensure prebuilt table is compatible 4) Looks for mer mask files based on mersize 5) Replaces system() with runCommand()

+++ kmer/ESTmapper/util/2-search.pl
@@ -1,28 +1,38 @@
 use strict;
 
 sub checkFinished {
-    my ($path, $cdnaInInput, @scafList) = @_;
+    my ($path, $cdnaInInput, $buildonly, $buildprefix, @scafList) = @_;
     my @searchesToRun;
 
     open(S, "> $path/1-search/run.sh");
     foreach my $s (@scafList) {
         my $c = 0;
 
-        #  If the hits file is NOT found, remove the count file.
-        #
-        unlink "$path/1-search/$s.count" if (! -e "$path/1-search/$s.hits");
+        if ($buildonly) {
 
-        #  If there is a count file, count the number of lines in it.
-        #
-        $c = int(`wc -l < $path/1-search/$s.count`) if (-e "$path/1-search/$s.count");
+            #  We're done if the table file exists for this piece
+            #
+            if (! -e "$buildprefix.$s") {
+                print S "$path/1-serach/$s.cmd";
+                push @searchesToRun, "$path/1-search/$s.cmd";
+            }
+        } else {
 
-        #  There should be one for each cDNA.
-        #
-        if ($c != $cdnaInInput) {
-            print STDERR "ESTmapper/search-- Found signal for $c cDNA out of $cdnaInInput for segment $s.\n";
+            #  If the hits file is NOT found, remove the count file.
+            #
+            unlink "$path/1-search/$s.count" if (! -e "$path/1-search/$s.hits");
 
-            print S "$path/1-search/$s.cmd\n";
-            push @searchesToRun, "$path/1-search/$s.cmd";
+            #  If there is a count file, count the number of lines in it.
+            #
+            $c = int(`wc -l < $path/1-search/$s.count`) if (-e "$path/1-search/$s.count");
+
+            #  There should be one for each cDNA.
+            #
+            if ($c != $cdnaInInput) {
+                print STDERR "ESTmapper/search-- Found signal for $c cDNA out of $cdnaInInput for segment $s.\n";
+                print S "$path/1-search/$s.cmd\n";
+                push @searchesToRun, "$path/1-search/$s.cmd";
+            }
         }
     }
     close(S);
@@ -37,7 +47,8 @@ sub search {
     my $path         = "";
     my $opts         = "";
     my $mersize      = 20;
-    my $maskFile     = "";
+    my $maskFile     = undef;
+    my $buildonly    = 0;
     my $builddir     = undef;
     my $buildprefix  = undef;
     my $verbose      = "";
@@ -81,7 +92,13 @@ sub search {
             $mersize = shift @ARGS;
         }
 
-        if ($arg eq "-prebuild") {
+        if ($arg eq "-buildtables") {
+            $buildonly   = 1;
+            $builddir    = shift @ARGS;
+            $buildprefix = shift @ARGS;
+        }
+        if ($arg eq "-usetables") {
+            $buildonly   = 0;
             $builddir    = shift @ARGS;
             $buildprefix = shift @ARGS;
         }
@@ -103,27 +120,35 @@ sub search {
 
         if ($arg eq "-species") {
             $species  = shift @ARGS;
-            $maskFile = "";
-            if ($species eq "none") {
-                $maskFile = "";
-            } elsif (($species eq "human") ||
-                     ($species eq "hum")) {
-                $maskFile = "$exechome/data/frequentMers-hum-20.fasta";
-            } elsif (($species eq "mouse") ||
-                     ($species eq "mus")) {
-                $maskFile = "$exechome/data/frequentMers-mus-20.fasta";
-            } elsif (($species eq "rattus") ||
-                     ($species eq "rat")) {
-                $maskFile = "$exechome/data/frequentMers-rat-20.fasta";
-            } else {
-                print STDERR "ESTmapper/search-- Unknown species '$species'.\n";
-                $maskFile = "";
-            }
+            $maskFile = undef;
         }
 
         #  Use carefully or not at all.
         $maskFile = shift @ARGS if ($arg eq "-maskmers");
     }
+
+    if (!defined($maskFile)) {
+        if ($species eq "none") {
+            $maskFile = "";
+        } elsif (($species eq "human") ||
+                 ($species eq "hum")) {
+            $maskFile = "$exechome/data/frequentMers-hum-$mersize.fasta";
+        } elsif (($species eq "mouse") ||
+                 ($species eq "mus")) {
+            $maskFile = "$exechome/data/frequentMers-mus-$mersize.fasta";
+        } elsif (($species eq "rattus") ||
+                 ($species eq "rat")) {
+            $maskFile = "$exechome/data/frequentMers-rat-$mersize.fasta";
+        } else {
+            die "ESTmapper/search-- Unknown species '$species'.\n";
+        }
+
+        if (($maskFile ne "") && (! -e "$maskFile")) {
+            print STDERR "ESTmapper/search-- Can't find mer mask file '$maskFile'.\n";
+            print STDERR "ESTmapper/search-- Try a different mersize.\n";
+        }
+    }
+
 
     #  If we're all done, skip the searches.
     #
@@ -135,9 +160,8 @@ sub search {
     ($path eq "")                              and die "ERROR: ESTmapper/search-- no directory given.\n";
     (! -d "$path")                             and die "ERROR: ESTmapper/search-- no directory '$path' found!\n";
     (! -f "$path/0-input/scaffolds-list")      and die "ERROR: ESTmapper/search-- no scaffolds-list?\n";
-    (($maskFile ne "") && (! -e "$maskFile"))  and die "ERROR: ESTmapper/search-- Can't find mer mask file '$maskFile'!\n";
 
-    system("mkdir $path/1-search")                  if (! -d "$path/1-search");
+    mkdir "$path/1-search" if (! -d "$path/1-search");
 
     my $cdnaInInput  = int(`$leaff -F $path/0-input/cDNA.fasta -d`);
 
@@ -148,30 +172,108 @@ sub search {
     close(F);
     chomp @scafList;
 
+    #  Make the directory for building tables, and create an info file in there.
+    #  If we're $buildonly, we have $buildprefix and $builddir defined.
+    #
+    if ($buildonly) {
+        mkdir $buildprefix if (! -d $buildprefix);
+
+        my $genomepath = readlink "$path/0-input/genomic.fasta";
+
+        open(F, "> $builddir/$buildprefix.info");
+        print F "$genomepath\n";
+
+        foreach my $s (@scafList) {
+            my $segment;
+
+            open(G, "< $path/0-input/scaffolds-$s");
+            while (<G>) {
+                chomp;
+                $segment .= $_;
+            }
+            close(G);
+            print F "$segment\n";
+        }
+        close(F);
+    }
+
+    #  If we're reading tables back, make sure that the info file is
+    #  in agreement with our configuration.
+    #
+    #  This is sensitive to moving the genome file around.
+    #
+    if (defined($buildprefix) && !$buildonly) {
+        if (! -e "$builddir/$buildprefix.info") {
+            print STDERR "ESTmapper/search-- Table incompatible: No info file found!\n";
+            exit(1);
+        }
+
+        open(F, "< $builddir/$buildprefix.info");
+
+        my $genomepath = readlink "$path/0-input/genomic.fasta";
+        my $genomesave = <F>;
+        chomp $genomesave;
+
+        if ($genomepath ne $genomesave) {
+            print STDERR "ESTmapper/search-- Table incompatible: built using $genomesave\n";
+            print STDERR "ESTmapper/search--                     this run is $genomepath\n";
+            exit(1);
+        }
+
+        foreach my $s (@scafList) {
+            my $segment;
+
+            if (! -e "$builddir/$buildprefix.$mersize.$s") {
+                print STDERR "ESTmapper/search-- Table incompatible: missing segment $s\n";
+                exit(1);
+            }
+
+            open(G, "< $path/0-input/scaffolds-$s");
+            while (<G>) {
+                chomp;
+                $segment .= $_;
+            }
+            close(G);
+
+            my $tablesegment = <F>;
+            chomp $tablesegment;
+
+            if ($segment ne $tablesegment) {
+                print STDERR "ESTmapper/search-- Table incompatible: segment $s is wrong.\n";
+                exit(1);
+            }
+        }
+        close(F);
+    }
+
     #  Create a bunch of scripts to process
     #
     #  Rewrite the command everytime.  This fixes the problem where
-    #  we would, say, change the stats or number of threads.
+    #  we would, say, change the stats or number of threads, and is
+    #  needed for switching from -buildtables to -usetables.
     #
     foreach my $s (@scafList) {
         open(F, "> $path/1-search/$s.cmd");
         print F "$searchGENOME $verbose -binary -mersize $mersize $opts -numthreads $numthread";
-        print F " -buildonly $builddir/$buildprefix.$mersize.$s" if (defined($buildprefix));
+        print F " -buildtables $builddir/$buildprefix.$mersize.$s" if (defined($buildprefix) &&  $buildonly);
+        print F " -usetables   $builddir/$buildprefix.$mersize.$s" if (defined($buildprefix) && !$buildonly);
         print F " -cdna $path/0-input/cDNA.fasta";
         print F " -genomic $path/0-input/genomic.fasta";
         print F " -use $path/0-input/scaffolds-$s";
         print F " -output $path/1-search/$s.hits";
         print F " -count $path/1-search/$s.count";
-        print F " -stats $path/1-search/$s.stats" if ($stats == 1);
+        print F " -stats $builddir/$buildprefix.$mersize.$s.stats" if (($stats == 1) &&  $buildonly);
+        print F " -stats $path/1-search/$s.stats"                  if (($stats == 1) && !$buildonly);
         print F " -mask $maskFile" if ($maskFile ne "");
         close(F);
-        system("chmod 755 $path/1-search/$s.cmd");
+
+        chmod 0755, "$path/1-search/$s.cmd";
     }
 
     #  We build a list of jobs to run, and put it into
     #  $path/1-search/run.sh.
     #
-    my @searchesToRun = checkFinished($path, $cdnaInInput, @scafList);
+    my @searchesToRun = checkFinished($path, $cdnaInInput, $buildonly, "$builddir/$buildprefix.$mersize", @scafList);
 
     #  Run searches.  If the search terminated properly, the
     #  hit-counts file should exist.  Run (maybe re-run) the search if
@@ -203,7 +305,7 @@ sub search {
                 #  See if anything failed.
                 #
                 print STDERR "ESTmapper/search-- checking search output.  All should have $cdnaInInput cDNA.\n";
-                @searchesToRun = checkFinished($path, $cdnaInInput, @scafList);
+                @searchesToRun = checkFinished($path, $cdnaInInput, $buildonly, "$builddir/$buildprefix.$mersize", @scafList);
             }
 
             if (($numTries < 3) && (scalar(@searchesToRun) > 0)) {
@@ -250,7 +352,9 @@ sub search {
                 $cmd .= " -J f$farmname ";
                 $cmd .= " $ESTmapper -restart $path";
 
-                system($cmd);
+                if (runCommand($cmd)) {
+                    die "Failed to submit job to LSF.\n";
+                }
 
                 print STDERR "ESTmapper/search-- Searches submitted.   Rest of run is on the farm.\n";
 
@@ -264,39 +368,64 @@ sub search {
     }
 
 
+    #  Summarize compute time used.
+    #
+    my $sysTimeBuild  = 0;
+    my $usrTimeBuild  = 0;
+    my $sysTimeSearch = 0;
+    my $usrTimeSearch = 0;
+    foreach my $s (@scafList) {
+        if (-e "$path/1-search/$s.stats") {
+            open(F, "< $path/1-search/$s.stats");
+            while (<F>) {
+                $sysTimeSearch += $1 if (m/^systemTime:\s+(\d+\.\d+)$/);
+                $usrTimeSearch += $1 if (m/^userTime:\s+(\d+\.\d+)$/);
+            }
+            close(F);
+        }
+        if (-e "$builddir/$buildprefix.$mersize.$s.stats") {
+            open(F, "< $builddir/$buildprefix.$mersize.$s.stats");
+            while (<F>) {
+                $sysTimeBuild += $1 if (m/^systemTime:\s+(\d+\.\d+)$/);
+                $usrTimeBuild += $1 if (m/^userTime:\s+(\d+\.\d+)$/);
+            }
+            close(F);
+        }
+    }
+
+    #  If we were supposed to be doing just table builds, we either crashed or are finished.
+    #  Either way, we exit.
+    #
+    if ($buildonly) {
+        print STDERR "ESTmapper/search-- Used $sysTimeBuild seconds system time to build tables.\n";
+        print STDERR "ESTmapper/search-- Used $usrTimeBuild seconds user time to build tables.\n";
+        print STDERR "ESTmapper/search-- Script finished in ", time() - $startTime, " wall-clock seconds.\n" if (time() > $startTime + 5);
+        print STDERR "ESTmapper/search-- All finished building tables.\n";
+        exit(0);
+    }
+
     #  See if anything is still broken.  If it did, holler!
     #  We also sum the time used here.
     #
-    my $died    = 0;
-    my $sysTime = 0;
-    my $usrTime = 0;
+    my $died = 0;
     foreach my $s (@scafList) {
         if (! -e "$path/1-search/$s.count") {
             print STDERR "ERROR: ESTmapper/search-- SEARCH $s DIED!\n";
             $died++;
-        }
-        if (-e "$path/1-search/$s.stats") {
-            open(F, "< $path/1-search/$s.stats");
-            while (<F>) {
-                if (m/^systemTime:\s+(\d+\.\d+)$/) {
-                    $sysTime += $1;
-                }
-                if (m/^userTime:\s+(\d+\.\d+)$/) {
-                    $usrTime += $1;
-                }
-            }
-            close(F);
         }
     }
     die "ERROR: ESTmapper/search-- Searches have a problem ($died).\n" if ($died > 0);
 
     #  Rather lazy way to inform the next step (and future calls to this step) that we're all done.
     #
-    system("touch $path/1-search/allDone");
+    open(F, "> $path/1-search/allDone");
+    close(F);
 
-    print STDERR "ESTmapper: searchGENOME required $sysTime seconds system time.\n";
-    print STDERR "ESTmapper: searchGENOME required $usrTime seconds user time.\n";
-    print STDERR "ESTmapper: Search script finished in ", time() - $startTime, " wall-clock seconds.\n" if (time() > $startTime + 5);
+    print STDERR "ESTmapper/search-- Used $sysTimeBuild seconds system time to precompute tables.\n";
+    print STDERR "ESTmapper/search-- Used $usrTimeBuild seconds user time to precompute tables.\n";
+    print STDERR "ESTmapper/search-- Used $sysTimeSearch seconds system time.\n";
+    print STDERR "ESTmapper/search-- Used $usrTimeSearch seconds user time.\n";
+    print STDERR "ESTmapper/search-- Script finished in ", time() - $startTime, " wall-clock seconds.\n" if (time() > $startTime + 5);
 }
